Added length option to settings modal
