+++ src/app/core/components/editor/modal-element-properties/modal-element-properties.component.html
@@ -19,11 +19,22 @@
       <input type="color" id="selectedColorPicker" [(ngModel)]="selectedColor" (change)="onSelectedColorChange()">
     </div>
 
-    <br>
+    <div *ngIf="isWall()">
+      <br>
+      <div class="thicknessSetting">
+        <label for="thickness">{{ "Thickness:" | translate }}</label>
+        <input type="range" id="thickness" name="thickness" min="5" max="20" step="0.1" [(ngModel)]="thickness"
+               (change)="onThicknessChange()"/>
+      </div>
+    </div>
 
-    <div class="thicknessSetting" *ngIf="isWall()">
-      <label for="thickness">{{ "Thickness:" | translate }}</label>
-      <input type="range" id="thickness" name="thickness" min="5" max="20" step="0.1" [(ngModel)]="thickness" (change)="onThicknessChange()" />
+    <div *ngIf="isWallElement()">
+      <br>
+      <div class="lengthSetting">
+        <label for="length">{{ "Height:" | translate }}</label>
+        <input type="range" id="length" name="length" min="5" max="20" step="0.1" [(ngModel)]="length"
+               (change)="onLengthChange()"/>
+      </div>
     </div>
   </div>
