+++ src/app/core/models/wall-elements/window.ts
@@ -163,20 +163,4 @@ export class Window extends WallElement {
   setVisibleState(newState: boolean): void {
   }
 
-  override getColor(): string | undefined {
-    return this.color;
-  }
-
-  override getSelectedColor(): string | undefined {
-    return this.selectedColor;
-  }
-
-  override setSelectedColor(color?: string): void {
-    this.selectedColor = color;
-  }
-
-  override setColor(newColor?: string) {
-    this.color = newColor;
-  }
-
 }
